% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrclm.R, R/zzz.R
\docType{package}
\name{mbrclm}
\alias{mbrclm}
\title{Median bias reduction in cumulative link models}
\usage{
mbrclm(
  x,
  y,
  start = NULL,
  type = c("AS_median", "AS_mean", "AS_ml"),
  link = c("logit", "probit", "cloglog"),
  maxiter = 100,
  tolerance = 10^(-10)
)
}
\arguments{
\item{x}{model matrix (without intercept)}

\item{y}{response matrix (each row with an indicator of observed category)}

\item{start}{vector of starting values for the algorithm}

\item{type}{type of adjustment of the score function}

\item{link}{link function}

\item{maxiter}{maximum number of iterations for the algorithm}

\item{tolerance}{value used for the stopping criterion}
}
\value{
Parameter estimates
}
\description{
Such a function contains all the needed quantities to fit the model

The mbrclm package contains all the needed quantities for fitting cumualtive link models via median bias reduction and get the summary
}
\section{Mypackage functions}{

mbrclm
summary.mbrclm
print.summary.mbrclm
}

