% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.summary.mbrclm}
\alias{print.summary.mbrclm}
\title{mbrclm: print summary}
\usage{
\method{print}{summary.mbrclm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{object of class mbrclm}

\item{digits}{number of digits of the printed values}

\item{...}{further arguments}
}
\value{
Summary print
}
\description{
Such a function allows for printing the results of the fit
}
