% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.mbrclm}
\title{mbrclm: get the summary}
\usage{
\method{summary}{mbrclm}(object, ...)
}
\arguments{
\item{object}{object of class mbrclm}

\item{...}{further arguments}
}
\value{
The summary is printed
}
\description{
Such a function allows for obtaining the results of the fit
}
