% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{World Health Organization TB data}
\format{
\subsection{\code{wine}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{temp}{Year}
\item{contact}{Year}
\item{rating_comb3}{Year}
...
}
}
}
\source{
from the sataset wine of the ordinal poackage
}
\usage{
wine
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
